+++ printing/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.7.3
 
 - Update pdfium version to 4861
+- Fix crash when Android load a PDF file which had password
 
 ## 5.7.2