+++ printing/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update pdfium version to 4861
 - Fix crash when Android load a PDF file which had password
 - Fix parsing TTF fonts with zero-length glyphs
+- Fix PdfPreview page format and orientation updates
 
 ## 5.7.2